refactor: one trusted publishing configuration class, not two
TrustedPublishingProperties existed only because @Value cannot convert a
Duration in the bare SpringExtension contexts the provider tests use, which
did not justify a second class reached through pass-through getters. Both
now live on TrustedPublishingConfig: the scalars stay on @Value, and only
the two typed settings have setters, so the configuration properties binder
touches exactly those. That keeps the audience fallback to ovsx.webui.url
expressible as a placeholder default, which the binder cannot do.

Also removes getAllTrustedPublisherProviders, which had no production
caller and was kept alive by an assertion in its own test, replaces a
verbose three-arg collect with Collectors.toMap, and states the created_by
foreign key the way its siblings in the migration now read: a registration
cannot outlive its author any more than it outlives their ownership.
