+++ server/src/test/java/org/eclipse/openvsx/LocalRegistryServiceTest.java
@@ -51,7 +51,6 @@ import org.eclipse.openvsx.search.SearchUtilService;
 import org.eclipse.openvsx.search.SimilarityCheckService;
 import org.eclipse.openvsx.storage.StorageUtilService;
 import org.eclipse.openvsx.trustedpublishing.TrustedPublishingConfig;
-import org.eclipse.openvsx.trustedpublishing.TrustedPublishingProperties;
 import org.eclipse.openvsx.util.ErrorResultException;
 import org.eclipse.openvsx.util.TempFile;
 import org.eclipse.openvsx.util.VersionService;
@@ -129,7 +128,7 @@ class LocalRegistryServiceTest {
                 integrityService,
                 similarityCheckService,
                 new PublishingConfig(),
-                new TrustedPublishingConfig(new TrustedPublishingProperties()),
+                new TrustedPublishingConfig(),
                 Duration.ofSeconds(30));
 
         // A permissive default for a void method rather than a per-test expectation: the tests of
