+++ server/src/test/java/org/eclipse/openvsx/RegistryAPITest.java
@@ -81,7 +81,6 @@ import org.eclipse.openvsx.security.SecurityConfig;
 import org.eclipse.openvsx.storage.*;
 import org.eclipse.openvsx.storage.log.DownloadCountService;
 import org.eclipse.openvsx.trustedpublishing.TrustedPublishingConfig;
-import org.eclipse.openvsx.trustedpublishing.TrustedPublishingProperties;
 import org.eclipse.openvsx.util.ChangesCursor;
 import org.eclipse.openvsx.util.LogService;
 import org.eclipse.openvsx.util.NamingUtil;
@@ -3751,7 +3750,7 @@ class RegistryAPITest {
 
         @Bean
         TrustedPublishingConfig trustedPublishingConfig() {
-            return new TrustedPublishingConfig(new TrustedPublishingProperties());
+            return new TrustedPublishingConfig();
         }
 
         @Bean
