+++ server/src/test/java/org/eclipse/openvsx/admin/AdminAPITest.java
@@ -104,7 +104,6 @@ import org.eclipse.openvsx.storage.LocalStorageService;
 import org.eclipse.openvsx.storage.StorageUtilService;
 import org.eclipse.openvsx.storage.log.DownloadCountService;
 import org.eclipse.openvsx.trustedpublishing.TrustedPublishingConfig;
-import org.eclipse.openvsx.trustedpublishing.TrustedPublishingProperties;
 import org.eclipse.openvsx.util.LogService;
 import org.eclipse.openvsx.util.TargetPlatform;
 import org.eclipse.openvsx.util.TargetPlatformVersion;
@@ -2654,7 +2653,7 @@ class AdminAPITest {
                     integrityService,
                     similarityCheckService,
                     new PublishingConfig(),
-                    new TrustedPublishingConfig(new TrustedPublishingProperties()),
+                    new TrustedPublishingConfig(),
                     Duration.ofSeconds(30));
         }
